requirements.txt
Django>=5.0,<5.2.8
djangorestframework>=3.15.0,<3.16.0
psycopg2-binary>=2.9.9,<2.9.11
